Completed Orders add to vuforia and gathering targets ration on Index page

+++ Glaz.Server/Data/Vuforia/Responses/TargetRecord.cs
@@ -15,6 +15,6 @@ namespace Glaz.Server.Data.Vuforia.Responses
         public float Width { get; set; }
 
         [JsonProperty("tracking_rating")]
-        public byte TrackingRating { get; set; }
+        public int TrackingRating { get; set; }
     }
 }
